Move runLocal and runLocal to /src/lib
